Refactor app to support multiple endpoints.
