+++ operator/keip/requirements-dev.txt
@@ -1,0 +1,6 @@
+black==25.1.0
+coverage==7.10.6
+httpx==0.28.1
+mypy==1.18.1
+pytest==8.4.2
+ruff==0.13.0
