+++ operator/keip/requirements.txt
@@ -1,0 +1,2 @@
+starlette==0.47.3
+uvicorn[standard]==0.35.0
