+++ operator/keip/routes/webhook.py
@@ -1,19 +1,17 @@
 import json
-import logging.config
+import logging
 from json import JSONDecodeError
 from typing import Callable, Mapping
 
-from starlette.applications import Starlette
 from starlette.exceptions import HTTPException
 from starlette.requests import Request
 from starlette.responses import JSONResponse
-from starlette.routing import Route
+from starlette.routing import Route, Router
 from starlette.status import HTTP_400_BAD_REQUEST
 
-from webhook import config as cfg
-from webhook.core.sync import sync
-from webhook.addons.certmanager.main import sync_certificate
-from webhook.logconf import LOG_CONF
+from core.sync import sync
+from addons.certmanager.main import sync_certificate
+
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -45,20 +43,14 @@ async def status(request):
     return JSONResponse({"status": "UP"})
 
 
-routes = [
-    Route("/sync", endpoint=build_webhook(sync), methods=["POST"]),
-    Route(
-        "/addons/certmanager/sync",
-        endpoint=build_webhook(sync_certificate),
-        methods=["POST"],
-    ),
-    Route("/status", endpoint=status, methods=["GET"]),
-]
-
-
-logging.config.dictConfig(LOG_CONF)
-
-if cfg.DEBUG:
-    _LOGGER.warning("Running server with debug mode. NOT SUITABLE FOR PRODUCTION!")
-
-app = Starlette(debug=cfg.DEBUG, routes=routes)
+router = Router(
+    [
+        Route("/sync", endpoint=build_webhook(sync), methods=["POST"]),
+        Route(
+            "/addons/certmanager/sync",
+            endpoint=build_webhook(sync_certificate),
+            methods=["POST"],
+        ),
+        Route("/status", endpoint=status, methods=["GET"]),
+    ]
+)
