--- operator/webhook/addons/certmanager/test/json/full-iroute-request.json
@@ -1,49 +1,0 @@
-{
-  "object": {
-    "apiVersion": "keip.codice.org/v1alpha1",
-    "kind": "IntegrationRoute",
-    "metadata": {
-      "annotations": {
-        "cert-manager.io/alt-names": "cloud-integration-route-actuator.testnamespace.svc.cluster.local",
-        "cert-manager.io/cluster-issuer": "test-selfsigned",
-        "cert-manager.io/common-name": "testroute",
-        "cert-manager.io/subject-countries": "US",
-        "cert-manager.io/subject-localities": "A Park",
-        "cert-manager.io/subject-organizationalunits": "Parks and Recreation",
-        "cert-manager.io/subject-provinces": "FL"
-      },
-      "creationTimestamp": "2025-03-18T15:55:04Z",
-      "generation": 1,
-      "name": "testroute",
-      "namespace": "testnamespace",
-      "resourceVersion": "48692499",
-      "uid": "c4d501ce-0196-42f1-bd8c-8dd82759d890"
-    },
-    "spec": {
-      "propSources": [
-        {
-          "name": "testroute-props"
-        }
-      ],
-      "replicas": 1,
-      "routeConfigMap": "testroute-xml",
-      "secretSources": [
-        "testroute-secret"
-      ],
-      "tls": {
-        "keystore": {
-          "jks": {
-            "key": "keystore.jks",
-            "passwordSecretRef": "jks-password",
-            "secretName": "testroute-certstore"
-          }
-        }
-      }
-    }
-  },
-  "attachments": {
-    "Certificate.cert-manager.io/v1": {}
-  },
-  "related": {},
-  "finalizing": false
-}
