--- operator/webhook/requirements-dev.txt
@@ -1,6 +1,0 @@
-black==25.1.0
-coverage==7.7.1
-httpx==0.28.1
-mypy==1.15.0
-pytest==8.3.5
-ruff==0.11.2
