--- operator/webhook/requirements.txt
@@ -1,2 +1,0 @@
-starlette==0.41.3
-uvicorn[standard]==0.29.0
