--- operator/webhook/test/json/full-cert-request.json
@@ -1,86 +1,0 @@
-{
-  "controller": {
-    "kind": "DecoratorController",
-    "apiVersion": "metacontroller.k8s.io/v1alpha1",
-    "metadata": {
-      "name": "keip-certmanager-controller",
-      "uid": "79dfff16-d192-439f-8af2-aaa75409514e",
-      "resourceVersion": "48689094",
-      "generation": 1,
-      "creationTimestamp": "2025-03-18T15:19:49Z"
-    },
-    "spec": {
-      "resources": [
-        {
-          "apiVersion": "keip.codice.org/v1alpha1",
-          "resource": "integrationroutes"
-        }
-      ],
-      "attachments": [
-        {
-          "apiVersion": "cert-manager.io/v1",
-          "resource": "certificates",
-          "updateStrategy": {
-            "method": "RollingRecreate"
-          }
-        }
-      ],
-      "hooks": {
-        "sync": {
-          "version": "v1",
-          "webhook": {
-            "url": "http://integrationroute-webhook.keip/addons/certmanager/sync"
-          }
-        }
-      }
-    },
-    "status": {}
-  },
-  "object": {
-    "apiVersion": "keip.codice.org/v1alpha1",
-    "kind": "IntegrationRoute",
-    "metadata": {
-      "annotations": {
-        "cert-manager.io/alt-names": "cloud-integration-route-actuator.testnamespace.svc.cluster.local",
-        "cert-manager.io/cluster-issuer": "test-selfsigned",
-        "cert-manager.io/common-name": "testroute",
-        "cert-manager.io/subject-countries": "US",
-        "cert-manager.io/subject-localities": "A Park",
-        "cert-manager.io/subject-organizationalunits": "Parks and Recreation",
-        "cert-manager.io/subject-provinces": "FL"
-      },
-      "creationTimestamp": "2025-03-18T15:55:04Z",
-      "generation": 1,
-      "name": "testroute",
-      "namespace": "testnamespace",
-      "resourceVersion": "48692499",
-      "uid": "c4d501ce-0196-42f1-bd8c-8dd82759d890"
-    },
-    "spec": {
-      "propSources": [
-        {
-          "name": "testroute-props"
-        }
-      ],
-      "replicas": 1,
-      "routeConfigMap": "testroute-xml",
-      "secretSources": [
-        "testroute-secret"
-      ],
-      "tls": {
-        "keystore": {
-          "jks": {
-            "key": "keystore.jks",
-            "passwordSecretRef": "jks-password",
-            "secretName": "testroute-certstore"
-          }
-        }
-      }
-    }
-  },
-  "attachments": {
-    "Certificate.cert-manager.io/v1": {}
-  },
-  "related": {},
-  "finalizing": false
-}
